refactor: move TeamFeedbackRow into own component
